下载歌曲接口上送id属性修改为长整形

+++ src/main/java/com/baltan/notease/music/service/SongServiceImpl.java
@@ -112,7 +112,7 @@ public class SongServiceImpl implements SongService {
         String responseMessage = Response.SUCCESSFUL.getMESSAGE();
 
         try {
-            int id = (int) params.get("id");
+            long id = (long) params.get("id");
             List<String> artistNames = (List<String>) params.get("artistNames");
             String songName = (String) params.get("songName");
             FileUtil.downloadSong(id, artistNames, songName);
